+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.2.0
+
+### Added
+
+- Adds command `Set-TeamViewerManagedDevice` to change properties of a managed device.
+- Adds optional `Policy` property to `Get-TeamViewerManagedDevice` result entries.
+- Adds `-Device` parameter to `Get-TeamViewerManagedGroup` command, allowing to
+  fetch all managed groups that a particular device is part of.
+
 ## 1.1.0 (2021-04-15)
 
 ### Added